Added hyperdrive, colors changed, error for non-pages
Implemented a new option in the Navigation menu to enter or exit hyperspace; this option only works if your ship fits the size requirement. Theme of the entire OS shifted to a dark mode. The broken Crew page, alongside the Target option in Navigation, now only has an option that gives an error that it doesn't work.