Added install feature
In the "Advanced" menu, there is now an option that installs the entire OS onto another drive on the system. This was added to facilitate installing the OS to a HDD from a floppy drive, which will be provided as loot in a pack I am assembling.
Any drive can be installed to. There is not an actual check for whether it is empty or not, and installing to a non-empty drive will probably break things.